pyproject.toml
[project]
name = "governed-financial-advisor"
version = "0.1.0"
description = "AI-driven agent designed to facilitate the exploration of the financial advisor landscape"
authors = [{ name = "Lars Ahlfors", email = "laah@google.com" }]
license = "Apache-2.0"
readme = "README.md"
dependencies = [
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk[extensions]>=0.11.2",
    "litellm>=1.0.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-requests>=0.58b0",
    "nemoguardrails>=0.17.0",
    "nest-asyncio>=1.6.0",
    "redis>=5.0.0",
    "langgraph<0.4.8",
    "langgraph-checkpoint-redis",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-langchain",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "httpx>=0.27.0",
    "python-json-logger>=2.0.7",
    "uvicorn>=0.29.0",
    "fastapi>=0.110.0",
    "openai>=1.0.0",
    "kfp>=2.6.0",
    "mcp>=0.1.0",
    "sse-starlette>=1.0.0",
    "grpcio>=1.78.0",
    "outlines",
]

requires-python = ">=3.10,<3.13"

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "nest-asyncio>=1.6.0",
    "respx>=0.22.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
]

deployment = [
    "absl-py>=2.2.1",
]

[project.optional-dependencies]

lint = [
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties and settings.
[tool.agent-starter-pack]
example_question = "Analyze AAPL"

[tool.setuptools.packages.find]
where = ["src"]

[tool.agent-starter-pack.settings]
agent_directory = "src/governed_financial_advisor/agents"

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
